Mods to add auto documentation generation with the petrbela/blueprint gem. Also combined all blueprint out files plus intro text into public/developer/api.md for conversion to html by Aglio.
