+++ api_docs/profile.md
@@ -1,0 +1,76 @@
+# Group Profile
+
+
+## GET /api/v1/profiles
+
+List all profiles.
+
++ Response 200 (application/json; charset=utf-8)
+
+    {
+      "title": null,
+      "first_name": "Joe",
+      "middle_name": null,
+      "last_name": null,
+      "suffix": null,
+      "address": null,
+      "address2": null,
+      "city": null,
+      "state": null,
+      "zip": null,
+      "gender": null,
+      "marital_status": null,
+      "is_parent": null,
+      "is_student": null,
+      "is_veteran": null,
+      "is_retired": null,
+      "email": null,
+      "phone_number": null,
+      "mobile_number": null,
+      "uid": "d553c6dd-e4cb-4827-8783-53dd65ff828e",
+      "id": "d553c6dd-e4cb-4827-8783-53dd65ff828e"
+    }
+
++ Response 200 (application/json; charset=utf-8)
+
+    {
+      "title": null,
+      "first_name": "Joe",
+      "middle_name": null,
+      "last_name": "Citizen",
+      "suffix": null,
+      "address": null,
+      "address2": null,
+      "city": null,
+      "state": null,
+      "zip": null,
+      "gender": null,
+      "marital_status": null,
+      "is_parent": null,
+      "is_student": null,
+      "is_veteran": null,
+      "is_retired": "0",
+      "email": "joe@citizen.org",
+      "phone_number": null,
+      "mobile_number": null,
+      "uid": "609b93c1-450f-406d-bd01-06a72044ee5f",
+      "id": "609b93c1-450f-406d-bd01-06a72044ee5f"
+    }
+
++ Response 200 (application/json; charset=utf-8)
+
+    {
+      "email": "joe@citizen.org",
+      "givenName": "Joe",
+      "additionalName": null,
+      "familyName": "Citizen",
+      "homeLocation": {
+        "streetAddress": "",
+        "addressLocality": null,
+        "addressRegion": null,
+        "postalCode": null
+      },
+      "telephone": null,
+      "gender": null
+    }
+
